requirements.txt
Django==3.1
Pillow
django-import-export
django-ckeditor
